push

+++ Zenith/Assets/Scripts/ATB_System.cs
@@ -1,16 +1,17 @@
 using UnityEngine;
 
+//harusnya ini ATB musuh yg barnya nda keliatan
 public class ATB_System : MonoBehaviour
 {
-    public int maxValue = 100;
-    public int currentValue = 0;
+    public float maxValue = 100;
+    public float currentValue = 0;
     void Start()
     {
-        
+
     }
 
     void Update()
     {
-        
+
     }
 }
